- Clean up some code/removed comments - Updated nuget packages - Updated to ASP.NET Core 2.1
