+++ source/DasBlog.Web.UI/TagHelpers/CategoriesListTagHelper.cs
@@ -1,10 +1,10 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using DasBlog.Core;
+﻿using DasBlog.Core;
 using DasBlog.Web.Models.BlogViewModels;
 using Microsoft.AspNetCore.Razor.TagHelpers;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
-namespace DasBlog.Web.Helpers
+namespace DasBlog.Web.TagHelpers
 {
 	public class CategoriesListTagHelper : TagHelper
 	{
